Replace --with-ketone with a --type option
The --with-ketone option has the somewhat undesirable property that it shows
both ketone and glucose readings at the same time. Add a --type option that can
be used to specify a sample type instead. The behaviour is the same as before if
the option isn't specified. When specified, it selects the data type to
display. The option may be specified multiple times.